fix(timestream/validator): relax depth tracking in WHERE conditions
It makes sense to check that the whole condition (e.g. `measure_name = `foo`)
is at the same depth; but it doesn't necessarily need to be at the same
depth as the original SELECT / WHERE: putting parentheses around
conditions in the WHERE cluase will increase the depth, but is a valid
case.